requirements.txt
google-cloud-vision
opencv-python
numpy
streamlit
pillow 